requirements.txt
pandas_datareader~=0.10.0
matplotlib~=3.4.3
python-dateutil~=2.8.2
plotly~=5.3.1
pandas~=1.3.3
numpy~=1.21.2
scipy~=1.7.1
scikit-learn~=1.0
requests~=2.26.0